# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClubHub API — multi-team soccer club management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people:
    get:
      operationId: listPeople
      tags: [people]
      summary: List all people in the club
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [manager, coach, player, parent]
      responses:
        "200":
          description: People
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
    post:
      operationId: createPerson
      tags: [people]
      summary: Create a person (player, parent, coach or manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /people/{personId}:
    get:
      operationId: getPerson
      tags: [people]
      summary: Get a person's full profile with memberships and family links
      parameters:
        - $ref: "#/components/parameters/PersonId"
      responses:
        "200":
          description: Person detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updatePerson
      tags: [people]
      summary: Update a person's profile
      parameters:
        - $ref: "#/components/parameters/PersonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          $ref: "#/components/responses/NotFound"
  /guardianships:
    post:
      operationId: createGuardianship
      tags: [people]
      summary: Link a guardian to a player's lifelong account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianshipInput"
      responses:
        "201":
          description: Linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardianship"
  /guardianships/{guardianshipId}:
    delete:
      operationId: removeGuardianship
      tags: [people]
      summary: Remove a family link
      parameters:
        - $ref: "#/components/parameters/GuardianshipId"
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
...
  /teams/{teamId}/posts:
    get:
      operationId: listTeamPosts
      tags: [feed]
      summary: List posts for a team
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [feed]
      summary: Post to a team
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [feed]
      summary: Get a post with its comments
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updatePost
      tags: [feed]
      summary: Edit or pin a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deletePost
      tags: [feed]
      summary: Delete a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /posts/{postId}/comments:
    post:
      operationId: addComment
      tags: [feed]
      summary: Comment on a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: integer
...
    PersonId:
      name: personId
      in: path
      required: true
      schema:
        type: integer
    GuardianshipId:
      name: guardianshipId
      in: path
      required: true
      schema:
        type: integer
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Person:
      type: object
      required: [id, firstName, lastName, fullName, isMinor]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        fullName: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        isMinor: { type: boolean }
        hasLogin: { type: boolean }
...
    TeamMembership:
      type: object
      required: [id, teamId, teamName, role]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        teamName: { type: string }
        role: { type: string, enum: [manager, coach, player] }
        jerseyNumber: { type: ["integer", "null"] }
        position: { type: ["string", "null"] }
...
    Guardianship:
      type: object
      required: [id, relationship, canManage, guardian, player]
      properties:
        id: { type: integer }
        relationship: { type: string, enum: [parent, guardian, other] }
        canManage: { type: boolean }
        guardian: { $ref: "#/components/schemas/Person" }
        player: { $ref: "#/components/schemas/Person" }
...
    PersonDetail:
      type: object
      required: [person, memberships, guardians, wards]
      properties:
        person: { $ref: "#/components/schemas/Person" }
        memberships:
          type: array
          items: { $ref: "#/components/schemas/TeamMembership" }
        guardians:
          type: array
          items: { $ref: "#/components/schemas/Guardianship" }
        wards:
          type: array
          items: { $ref: "#/components/schemas/Guardianship" }
    Post:
      type: object
      required: [id, teamId, teamName, body, pinned, createdAt, author, commentCount]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        teamName: { type: string }
        title: { type: ["string", "null"] }
        body: { type: string }
        pinned: { type: boolean }
        createdAt: { type: string }
        author: { $ref: "#/components/schemas/Person" }
        commentCount: { type: integer }
    PostDetail:
      type: object
      required: [post, comments]
      properties:
        post: { $ref: "#/components/schemas/Post" }
        comments:
          type: array
          items: { $ref: "#/components/schemas/Comment" }
    Comment:
      type: object
      required: [id, postId, body, createdAt, author]
      properties:
        id: { type: integer }
        postId: { type: integer }
        body: { type: string }
        createdAt: { type: string }
        author: { $ref: "#/components/schemas/Person" }
...
    PersonInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        dateOfBirth: { type: string }
    PersonUpdate:
      type: object
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        dateOfBirth: { type: string }
        avatarUrl: { type: string }
    GuardianshipInput:
      type: object
      required: [guardianId, playerId, relationship]
      properties:
        guardianId: { type: integer }
        playerId: { type: integer }
        relationship: { type: string, enum: [parent, guardian, other] }
        canManage: { type: boolean }
    PostInput:
      type: object
      required: [body]
      properties:
        title: { type: string }
        body: { type: string, minLength: 1 }
        pinned: { type: boolean }
    PostUpdate:
      type: object
      properties:
        title: { type: string }
        body: { type: string, minLength: 1 }
        pinned: { type: boolean }
    CommentInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
...